refactor: import `Variant` from `types.ts`
